refactor(deploy): extract shared copyDirForDeploy and rewriteHookPaths (DRY-4)
Create deploy-helpers.ts with shared implementations previously duplicated
in init.ts and deploy-global.ts. Both consumers now import from the shared
module. Preserves SEC-008 symlink traversal guard and keeps dryRun guard
at call sites in deploy-global.ts.
