+++ .planning/STATE.md
@@ -53,4 +53,4 @@ _State updated: 2026-03-14 — v4.4.0 milestone archived, state reset for next m
 
 ---
 
-_State generated from machine snapshot at 2026-03-16T18:41:59.315Z_
+_State generated from machine snapshot at 2026-03-16T18:52:04.879Z_
