+++ packages/luca-framework/src/utils/deploy-helpers.ts
@@ -1,0 +1,183 @@
+/**
+ * Shared deploy utility functions used by both `init.ts` and `deploy-global.ts`.
+ *
+ * Extracted from duplicate implementations (DRY-4) to provide a single source
+ * of truth for directory copying and hook path rewriting during deployment.
+ *
+ * Both consumers (`init.ts` and `scripts/deploy-global.ts`) previously had
+ * near-identical implementations of these functions. This module consolidates
+ * them into a single, shared API.
+ *
+ * **Security:** Preserves SEC-008 symlink traversal guard in `copyDirForDeploy`.
+ *
+ * @see packages/luca-framework/src/commands/init.ts — consumer (init flow)
+ * @see scripts/deploy-global.ts — consumer (global deploy script)
+ *
+ * @example
+ * ```typescript
+ * import { copyDirForDeploy, rewriteHookPaths } from "../utils/deploy-helpers";
+ *
+ * // Copy a directory for deployment, tracking files for manifest
+ * const deployedFiles: DeployedFileEntry[] = [];
+ * copyDirForDeploy(sourceDir, targetDir, globalDir, deployedFiles, "skill");
+ *
+ * // Rewrite relative paths in shell wrappers for global context
+ * // NOTE: callers must guard with their own dryRun check
+ * rewriteHookPaths(targetPath, projectRoot);
+ * ```
+ */
+import {
+  lstatSync,
+  mkdirSync,
+  readdirSync,
+  readFileSync,
+  realpathSync,
+  writeFileSync,
+} from "node:fs";
+import { join, relative } from "pathe";
+
+import type { DeploySourceType } from "./deploy-manifest.schemas";
+
+// ─── Types ──────────────────────────────────────────────────────────────────
+
+/**
+ * Entry representing a single deployed file, used for manifest tracking.
+ *
+ * Consumers pass an array of these entries to `copyDirForDeploy`, which
+ * appends to it as files are copied.
+ */
+export interface DeployedFileEntry {
+  /** Path relative to the global deploy directory (~/.claude/). */
+  relativePath: string;
+  /** Absolute path to the deployed file on disk. */
+  absolutePath: string;
+  /** Classification of the artifact source. */
+  sourceType: DeploySourceType;
+}
+
+// ─── Functions ──────────────────────────────────────────────────────────────
+
+/**
+ * Recursively copy a directory, tracking deployed files for manifest.
+ *
+ * Includes a symlink traversal guard (SEC-008): before copying each file,
+ * checks if it is a symlink. If the symlink resolves to a path outside
+ * the source tree, the file is skipped with a warning to prevent directory
+ * escape attacks during deployment.
+ *
+ * @param source - Absolute path to the source directory to copy from
+ * @param target - Absolute path to the target directory to copy into
+ * @param globalDir - Absolute path to the global deploy root (~/.claude/) for relative path computation
+ * @param deployedFiles - Mutable array to append deployed file entries to
+ * @param sourceType - Classification tag for all files copied from this directory
+ * @param sourceRoot - Internal: top-level source root for symlink validation (defaults to `source`)
+ *
+ * @example
+ * ```typescript
+ * const deployed: DeployedFileEntry[] = [];
+ * copyDirForDeploy(
+ *   "/path/to/.claude/skills/my-skill",
+ *   "/Users/me/.claude/skills/my-skill",
+ *   "/Users/me/.claude",
+ *   deployed,
+ *   "skill",
+ * );
+ * // deployed now contains entries for each file copied
+ * ```
+ */
+export function copyDirForDeploy(
+  source: string,
+  target: string,
+  globalDir: string,
+  deployedFiles: DeployedFileEntry[],
+  sourceType: DeploySourceType,
+  sourceRoot?: string,
+): void {
+  // Track the top-level source root for symlink validation
+  const root = sourceRoot ?? source;
+
+  mkdirSync(target, { recursive: true });
+  const entries = readdirSync(source, { withFileTypes: true });
+  for (const entry of entries) {
+    const srcPath = join(source, entry.name);
+    const tgtPath = join(target, entry.name);
+
+    // Symlink traversal guard (SEC-008)
+    if (lstatSync(srcPath).isSymbolicLink()) {
+      try {
+        const resolved = realpathSync(srcPath);
+        if (!resolved.startsWith(root)) {
+          console.warn(
+            `[deploy] Skipping symlink that escapes source tree: ${srcPath} -> ${resolved}`,
+          );
+          continue;
+        }
+      } catch {
+        console.warn(`[deploy] Skipping unresolvable symlink: ${srcPath}`);
+        continue;
+      }
+    }
+
+    if (entry.isDirectory()) {
+      copyDirForDeploy(
+        srcPath,
+        tgtPath,
+        globalDir,
+        deployedFiles,
+        sourceType,
+        root,
+      );
+    } else {
+      writeFileSync(tgtPath, readFileSync(srcPath));
+      deployedFiles.push({
+        relativePath: relative(globalDir, tgtPath),
+        absolutePath: tgtPath,
+        sourceType,
+      });
+    }
+  }
+}
+
+/**
+ * Rewrite relative paths in shell wrappers to absolute paths for global context.
+ *
+ * Shell wrappers generated by `generate-shell-wrappers.ts` use
+ * `$(dirname "$0")/../../src/hooks/scripts/` (hooks) or
+ * `$(dirname "$0")/../src/hooks/scripts/` (statusline). These resolve correctly
+ * inside the monorepo but break when the wrapper is copied to `~/.claude/`.
+ *
+ * This function replaces those relative prefixes with the absolute `projectRoot`
+ * path so the wrapper can invoke the TypeScript implementation from any location.
+ *
+ * **IMPORTANT:** The `../../` pattern is replaced before `../` to avoid partial
+ * matches (both patterns start with `$(dirname "$0")/../`).
+ *
+ * **NOTE:** This function does NOT include a dryRun guard. Callers are
+ * responsible for skipping the call in dry-run mode:
+ * ```typescript
+ * if (!dryRun) rewriteHookPaths(targetPath, projectRoot);
+ * ```
+ *
+ * @param targetPath - Absolute path to the deployed .sh wrapper file
+ * @param projectRoot - Absolute path to the monorepo root
+ *
+ * @example
+ * ```typescript
+ * // After copying a hook script to ~/.claude/hooks/
+ * rewriteHookPaths("/Users/me/.claude/hooks/session-start.sh", "/Users/me/code/luca-framework");
+ * // Replaces: $(dirname "$0")/../../ -> /Users/me/code/luca-framework/
+ * ```
+ */
+export function rewriteHookPaths(
+  targetPath: string,
+  projectRoot: string,
+): void {
+  const content = readFileSync(targetPath, "utf-8");
+  const rewritten = content
+    .replace(/\$\(dirname "\$0"\)\/\.\.\/\.\.\//g, `${projectRoot}/`)
+    .replace(/\$\(dirname "\$0"\)\/\.\.\//g, `${projectRoot}/`);
+
+  if (rewritten !== content) {
+    writeFileSync(targetPath, rewritten);
+  }
+}
