+++ scripts/deploy-global.ts
@@ -29,7 +29,6 @@ import {
   mkdirSync,
   readdirSync,
   readFileSync,
-  realpathSync,
   rmSync,
   symlinkSync,
   unlinkSync,
@@ -64,6 +63,12 @@ import {
 import { resolveCanonicalRegistry } from "../src/hooks/__helpers/hook-registry";
 import { generateClaudeHooksConfigFromCanonical } from "../src/hooks/__helpers/config-generators";
 
+import {
+  copyDirForDeploy,
+  rewriteHookPaths,
+} from "../packages/luca-framework/src/utils/deploy-helpers";
+
+import type { DeployedFileEntry } from "../packages/luca-framework/src/utils/deploy-helpers";
 import type { DeploySourceType } from "../packages/luca-framework/src/utils/deploy-manifest.schemas";
 
 // ─── Constants ──────────────────────────────────────────────────────────────
@@ -110,35 +115,8 @@ const log = (msg: string) => console.log(`  ${msg}`);
 const logHeader = (msg: string) => console.log(`\n→ ${msg}`);
 const logDry = (msg: string) => console.log(`  [dry-run] ${msg}`);
 
-/**
- * Rewrites relative paths in a deployed shell wrapper to absolute monorepo paths.
- *
- * Shell wrappers generated by generate-shell-wrappers.ts use
- * `$(dirname "$0")/../../src/hooks/scripts/` (hooks) or
- * `$(dirname "$0")/../src/hooks/scripts/` (statusline). These resolve correctly
- * inside the monorepo but break when the wrapper is copied to ~/.claude/.
- *
- * This function replaces those relative prefixes with the absolute `projectRoot`
- * path so the wrapper can invoke the TypeScript implementation from any location.
- *
- * IMPORTANT: The `../../` pattern is replaced before `../` to avoid partial
- * matches (both patterns start with `$(dirname "$0")/../`).
- *
- * @param targetPath - Absolute path to the deployed .sh wrapper file
- * @param projectRoot - Absolute path to the monorepo root
- */
-function rewriteWrapperPaths(targetPath: string, projectRoot: string): void {
-  if (dryRun) return;
-
-  const content = readFileSync(targetPath, "utf-8");
-  const rewritten = content
-    .replace(/\$\(dirname "\$0"\)\/\.\.\/\.\.\//g, `${projectRoot}/`)
-    .replace(/\$\(dirname "\$0"\)\/\.\.\//g, `${projectRoot}/`);
-
-  if (rewritten !== content) {
-    writeFileSync(targetPath, rewritten);
-  }
-}
+// rewriteWrapperPaths is now imported from deploy-helpers.ts
+// CRITICAL: callers must guard with dryRun check: `if (!dryRun) rewriteHookPaths(...)`
 
 /**
  * Symlink or copy a file. In copy mode, always copies.
@@ -227,65 +205,17 @@ function deployDir(
     symlinkSync(source, target);
     log(`symlink dir: ${basename(target)}/`);
   } else {
-    // Copy directory recursively
-    copyDirRecursive(source, target, sourceType);
+    // Copy directory recursively using shared deploy helper
+    if (sourceType) {
+      copyDirForDeploy(source, target, GLOBAL_DIR, deployedFiles, sourceType);
+    } else {
+      copyDirForDeploy(source, target, GLOBAL_DIR, deployedFiles, "lib");
+    }
     log(`copy dir: ${basename(target)}/`);
   }
 }
 
-/**
- * Recursively copy a directory, tracking files for manifest.
- *
- * Includes a symlink traversal guard (SEC-008): before copying each file,
- * checks if it is a symlink. If the symlink resolves to a path outside
- * the source tree, the file is skipped with a warning to prevent directory
- * escape attacks during deployment.
- */
-function copyDirRecursive(
-  source: string,
-  target: string,
-  sourceType?: DeploySourceType,
-  sourceRoot?: string,
-): void {
-  // Track the top-level source root for symlink validation
-  const root = sourceRoot ?? source;
-
-  mkdirSync(target, { recursive: true });
-  const entries = readdirSync(source, { withFileTypes: true });
-  for (const entry of entries) {
-    const srcPath = join(source, entry.name);
-    const tgtPath = join(target, entry.name);
-
-    // Symlink traversal guard (SEC-008)
-    if (lstatSync(srcPath).isSymbolicLink()) {
-      try {
-        const resolved = realpathSync(srcPath);
-        if (!resolved.startsWith(root)) {
-          console.warn(
-            `[deploy] Skipping symlink that escapes source tree: ${srcPath} -> ${resolved}`,
-          );
-          continue;
-        }
-      } catch {
-        console.warn(`[deploy] Skipping unresolvable symlink: ${srcPath}`);
-        continue;
-      }
-    }
-
-    if (entry.isDirectory()) {
-      copyDirRecursive(srcPath, tgtPath, sourceType, root);
-    } else {
-      writeFileSync(tgtPath, readFileSync(srcPath));
-      if (sourceType) {
-        deployedFiles.push({
-          relativePath: relative(GLOBAL_DIR, tgtPath),
-          absolutePath: tgtPath,
-          sourceType,
-        });
-      }
-    }
-  }
-}
+// copyDirRecursive is now replaced by copyDirForDeploy from deploy-helpers.ts
 
 // ─── Pre-flight Checks ─────────────────────────────────────────────────────
 
@@ -510,7 +440,7 @@ async function deployHooks(projectRoot: string): Promise<number> {
     }
 
     // Rewrite relative $(dirname "$0")/../../ paths to absolute monorepo paths
-    rewriteWrapperPaths(targetPath, projectRoot);
+    if (!dryRun) rewriteHookPaths(targetPath, projectRoot);
 
     count++;
   }
@@ -553,7 +483,7 @@ async function deployStatusline(projectRoot: string): Promise<void> {
   }
 
   // Rewrite relative $(dirname "$0")/../ paths to absolute monorepo paths
-  rewriteWrapperPaths(target, projectRoot);
+  if (!dryRun) rewriteHookPaths(target, projectRoot);
 
   log("statusline.sh deployed");
 }
